requirements.txt
annotated-types==0.6.0
anyio==4.3.0
badges @ git+https://github.com/EntySec/Badges@8767dca15b6c606935f22ffb5901b4100831e92b
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.6.2
bs4==0.0.2
camover @ git+https://github.com/EntySec/CamOver@478c4f3255de8d6582dfc0658461fa14c5e0ce1b
certifi==2022.12.7
cffi==1.15.1
cfscrape==2.1.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.6
click-plugins==1.1.1
cloudscraper==1.2.71
colorama==0.4.6
colorscript @ git+https://github.com/EntySec/ColorScript@b6f59d4e068832e60d52960d898186ed2d63a5b6
cryptography==41.0.3
dnspython==2.2.1
fastapi==0.110.3
filelock==3.13.4
Flask==1.1.2
future==0.18.3
h11==0.14.0
httptools==0.6.1
icmplib==3.0.3
idna==2.10
impacket==0.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
ldap3==2.9.1
ldapdomaindump==0.9.4
libretranslatepy==2.1.1
lxml==5.2.1
MarkupSafe==2.1.3
maxminddb==2.4.0
multidict==6.0.4
paramiko==3.4.0
psutil==5.9.5
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.7.1
pydantic_core==2.18.2
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyRoxy @ git+https://github.com/MHProDev/PyRoxy.git@ea0f88dbc0573292ba5672124f69e4e7a31b544d
PySocks==1.7.1
python-dotenv==1.0.1
python-nmap==0.6.1
PyYAML==6.0.1
requests==2.27.1
requests-file==2.0.0
requests-toolbelt==1.0.0
scapy==2.5.0
shodan==1.31.0
six==1.16.0
sniffio==1.3.1
socks==0
soupsieve==2.5
starlette==0.37.2
termcolor==2.3.0
tldextract==5.1.2
translate==3.6.1
typing_extensions==4.11.0
urllib3==1.25.10
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==2.3.6
XlsxWriter==3.2.0
yarl==1.8.2
